# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
/**
 * Original Tracks MP3 audio bytes
 */
export type MusicAudioFile = string;

export interface MusicProviderStatus {
  connected: boolean;
  label: string;
}

export type MusicGenerationInputMode = typeof MusicGenerationInputMode[keyof typeof MusicGenerationInputMode];


export const MusicGenerationInputMode = {
  instrumental: 'instrumental',
  vocals: 'vocals',
  demo: 'demo',
  loop: 'loop',
} as const;

export interface MusicGenerationInput {
  /**
     * @minLength 5
     * @maxLength 1500
     */
  prompt: string;
  mode: MusicGenerationInputMode;
  /**
     * @minimum 3
     * @maximum 600
     */
  durationSeconds: number;
  /** @maxLength 120 */
  genre?: string;
  /** @maxLength 120 */
  mood?: string;
  /** @maxLength 5000 */
  lyrics?: string;
  /** @maxLength 1000 */
  notes?: string;
}

export interface MusicGenerationResult {
  audioId: string;
  durationSeconds: number;
  contentType: string;
  /** Opaque one-time token for confirming or discarding this preview */
  previewToken: string;
}

export interface MusicPreviewAction {
  /**
     * @minLength 32
     * @maxLength 128
     */
  previewToken: string;
}

export interface MusicLyricsInput {
  /**
     * @minLength 5
     * @maxLength 1500
     */
  prompt: string;
  /** @maxLength 120 */
  genre?: string;
  /** @maxLength 120 */
  mood?: string;
  /** @maxLength 1000 */
  context?: string;
}

export interface MusicLyricsResult {
  lyrics: string;
}

/**
 * Raw audio bytes, up to 15 MB
 */
export type VoiceAudioFile = string;

export type VoiceTranscriptInputLanguage = typeof VoiceTranscriptInputLanguage[keyof typeof VoiceTranscriptInputLanguage];


export const VoiceTranscriptInputLanguage = {
  English: 'English',
  Korean: 'Korean',
  Mixed: 'Mixed',
} as const;

export interface VoiceTranscriptInput {
  /**
     * @minLength 3
     * @maxLength 2000
     */
  prompt: string;
  language: VoiceTranscriptInputLanguage;
}

export type VoiceTranscriptResultLanguage = typeof VoiceTranscriptResultLanguage[keyof typeof VoiceTranscriptResultLanguage];


export const VoiceTranscriptResultLanguage = {
  English: 'English',
  Korean: 'Korean',
  Mixed: 'Mixed',
} as const;

export interface VoiceTranscriptResult {
  title: string;
  transcript: string;
  context: string;
  delivery: string;
  category: string;
  relatedEvent?: string;
  language: VoiceTranscriptResultLanguage;
}

export type VoiceAudioInputVoice = typeof VoiceAudioInputVoice[keyof typeof VoiceAudioInputVoice];


export const VoiceAudioInputVoice = {
  alloy: 'alloy',
  echo: 'echo',
  onyx: 'onyx',
} as const;

export type VoiceAudioInputLanguage = typeof VoiceAudioInputLanguage[keyof typeof VoiceAudioInputLanguage];


export const VoiceAudioInputLanguage = {
  English: 'English',
  Korean: 'Korean',
  Mixed: 'Mixed',
} as const;

export interface VoiceAudioInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  transcript: string;
  voice: VoiceAudioInputVoice;
  language: VoiceAudioInputLanguage;
  /** @maxLength 300 */
  delivery?: string;
}

export interface VoiceAudioResult {
  audioId: string;
  durationSeconds: number;
}

export interface HealthStatus {
  status: string;
}

export interface ProviderStatusItem {
  connected: boolean;
  label: string;
}

export interface ProviderStatus {
  text: ProviderStatusItem;
  image: ProviderStatusItem;
  music: ProviderStatusItem;
  audio: ProviderStatusItem;
}

export type StoryGenerationInputMode = typeof StoryGenerationInputMode[keyof typeof StoryGenerationInputMode];


export const StoryGenerationInputMode = {
  story_update: 'story_update',
  generate_day: 'generate_day',
} as const;

export type StoryGenerationInputCanon = { [key: string]: unknown };

export type StoryGenerationInputExistingPhone = { [key: string]: unknown };

export interface StoryGenerationInput {
  /**
     * @minLength 3
     * @maxLength 12000
     */
  storyUpdate: string;
  mode: StoryGenerationInputMode;
  canon: StoryGenerationInputCanon;
  existingPhone: StoryGenerationInputExistingPhone;
}

export type ActivityProposalApp = typeof ActivityProposalApp[keyof typeof ActivityProposalApp];


export const ActivityProposalApp = {
  messages: 'messages',
  instagram: 'instagram',
  echo: 'echo',
  gallery: 'gallery',
  diary: 'diary',
  notes: 'notes',
  quickNotes: 'quickNotes',
  voice: 'voice',
  music: 'music',
  studio: 'studio',
  calendar: 'calendar',
  browser: 'browser',
  calls: 'calls',
  notifications: 'notifications',
  files: 'files',
  places: 'places',
  contacts: 'contacts',
} as const;

export type ActivityProposalMetadata = { [key: string]: unknown };

export interface ActivityProposal {
  id: string;
  app: ActivityProposalApp;
  type: string;
  title: string;
  content: string;
  timestamp: string;
  /** @nullable */
  person?: string | null;
  metadata: ActivityProposalMetadata;
}

export interface StoryGenerationResult {
  summary: string;
  proposals: ActivityProposal[];
}

export interface StoryImageInput {
  /**
     * @minLength 3
     * @maxLength 2000
     */
  prompt: string;
}

export interface StoryImageResult {
  dataUrl: string;
  caption: string;
  imageId?: string;
}

/**
 * Raw JPEG, PNG, or WebP bytes, up to 12 MB
 */
export type StoryImageFile = string;

export interface StoryImageUploadResult {
  imageId: string;
}

export type ConfirmOriginalTrackPreview200 = {
  confirmed: boolean;
};