# lib/api-zod/src/generated/types/musicAudioFile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Original Tracks MP3 audio bytes
 */
export type MusicAudioFile = string;