# lib/api-zod/src/generated/types/musicGenerationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MusicGenerationInputMode } from './musicGenerationInputMode';

export interface MusicGenerationInput {
  /**
     * @minLength 5
     * @maxLength 1500
     */
  prompt: string;
  mode: MusicGenerationInputMode;
  /**
     * @minimum 3
     * @maximum 600
     */
  durationSeconds: number;
  /** @maxLength 120 */
  genre?: string;
  /** @maxLength 120 */
  mood?: string;
  /** @maxLength 5000 */
  lyrics?: string;
  /** @maxLength 1000 */
  notes?: string;
}