# lib/api-zod/src/generated/types/musicGenerationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MusicGenerationResult {
  audioId: string;
  durationSeconds: number;
  contentType: string;
  /** Opaque one-time token for confirming or discarding this preview */
  previewToken: string;
}