# lib/api-zod/src/generated/types/musicLyricsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MusicLyricsInput {
  /**
     * @minLength 5
     * @maxLength 1500
     */
  prompt: string;
  /** @maxLength 120 */
  genre?: string;
  /** @maxLength 120 */
  mood?: string;
  /** @maxLength 1000 */
  context?: string;
}