# lib/api-zod/src/generated/types/musicPreviewAction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MusicPreviewAction {
  /**
     * @minLength 32
     * @maxLength 128
     */
  previewToken: string;
}